Add tests and loading window for eLL
