+++ PowerPointLabs/Test/Test.csproj
@@ -73,6 +73,7 @@
     <Reference Include="PresentationFramework" />
     <Reference Include="System" />
     <Reference Include="System.Drawing" />
+    <Reference Include="System.Net.Http" />
     <Reference Include="System.Speech" />
     <Reference Include="System.Windows.Forms" />
     <Reference Include="System.XML" />
@@ -136,6 +137,7 @@
     <Compile Include="UnitTest\ELearningLab\Model\ClickItemTest.cs" />
     <Compile Include="UnitTest\ELearningLab\Model\ComputerVoiceTest.cs" />
     <Compile Include="UnitTest\ELearningLab\Model\SelfExplanationClickItemTest.cs" />
+    <Compile Include="UnitTest\ELearningLab\Service\ELearningServiceTest.cs" />
     <Compile Include="UnitTest\FitToSlideTest.cs" />
     <Compile Include="UnitTest\HighlightLab\RemoveHighlightTest.cs" />
     <Compile Include="UnitTest\PictureSlidesLab\ModelFactory\StyleOptionsFactoryTest.cs" />
@@ -223,7 +225,6 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="UnitTest\ELearningLab\Service\" />
     <Folder Include="UnitTest\ELearningLab\Util\" />
   </ItemGroup>
   <Choose>
